Some changes added
Default images have been added to the register, the Profilecard component onClick function has been refactored to be a prop so that the Component can be used on Register. Images can now be previewed, but Login backend has not yet happened. Functionallity on Login and Register is essentially done from a frontend perpspective.